Moved core hello and by to core greet
