--- src/core-bye/CoreByeElement.html
@@ -1,4 +1,0 @@
-<label>
-    <span id="bye">Bye</span>
-    <slot id="name"></slot>
-</label>
