+++ src/core-greet/CoreGreetElement.html
@@ -1,4 +1,4 @@
 <label>
-    <span id="hello">Hello</span>
+    <span id="greet">Greet</span>
     <slot id="name"></slot>
 </label>
