+++ src/core-greet/CoreGreetElement.js
@@ -1,41 +1,33 @@
 import CoreElement from '../core-element/CoreElement';
 
-import TEMPLATE from './CoreByeElement.html';
-import STYLE from './CoreByeElement.css';
+import TEMPLATE from './CoreGreetElement.html';
+import STYLE from './CoreGreetElement.css';
 
 const STYLED_TEMPLATE = CoreElement.template(TEMPLATE, STYLE);
 
-const BYE_STRING = {
-  en: 'Bye',
-  es: 'Adios',
-  jp: 'さようなら',
-  fr: 'Au revoir',
-};
-
 /**
- * An element that displays "Bye" with the passed-in name.
+ * An element that displays some greeting from the language mapping with the passed-in name.
  * @property {string} name the string that will be displayed.
  * @property {boolean} rainbow whether the name should be colorfully animated.
  * @property {String} lang the language to display in.
  */
-class CoreByeElement extends CoreElement {
+class CoreGreetElement extends CoreElement {
   /**
-   * Creates a CoreBye element and attaches the shadow root
+   * Creates a CoreGreet element and attaches the shadow root
    */
-  constructor() {
+  constructor(langMap = null) {
     super(STYLED_TEMPLATE);
 
-    this.byeElement = this.shadowRoot.querySelector('#bye');
+    this.langMapping = langMap || { en: 'Greet' };
+
+    this.greetElement = this.shadowRoot.querySelector('#greet');
     this.nameElement = this.shadowRoot.querySelector('#name');
   }
 
   /** @private */
   static get properties() {
     return {
-      name: {
-        type: String,
-        value: 'World',
-      },
+      name: { type: String, value: 'World' },
       rainbow: { type: Boolean },
       lang: { type: String },
     };
@@ -48,13 +40,13 @@ class CoreByeElement extends CoreElement {
         this.nameElement.textContent = newValue;
         break;
       case 'lang':
-        this.byeElement.textContent = BYE_STRING[newValue] || BYE_STRING.en;
+        this.greetElement.textContent = this.langMapping[newValue] || this.langMapping.en;
         break;
       default:
     }
   }
 }
 
-CoreElement.customTag('core-bye', CoreByeElement);
+CoreElement.customTag('core-greet', CoreGreetElement);
 
-export default CoreByeElement;
+export default CoreGreetElement;
